Add files via upload
The motherboard of the rover houses the Primary ESP board, which includes the Central ESP (responsible for controlling the wheels, I2C lines, and a few RM motors) along with the Module ESP that handles additional GPIO outputs and some inputs. It also includes the Auxiliary ESP board, which provides extended control pins routed to a female header, allowing for future upgrades or expansions.

In addition, the motherboard hosts the Comms childboard, with a provision to toggle between an L2U (LAN to UART) module and a LoRa module (currently under integration). Another childboard is dedicated to the Raspberry Pi, which can take over wheel control from the ESPs depending on mission requirements. This Pi board also acts as a BNO node for inverse and forward kinematics (IK/FK) calculations.

The entire system is powered by a 12V supply from the Power Distribution Board (PDB).